# swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Reservas de Hotel API",
        "version": "1.0.0",
        "description": "API para gestionar reservas en hoteles"
    },
    "servers": [
        {
            "url": "http://localhost:3000/api"
        }
    ],
    "paths": {
        "/reservas": {
            "post": {
                "summary": "Crear una nueva reserva",
                "operationId": "crearReserva",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Reserva"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reserva creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Reserva"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en la solicitud"
                    }
                }
            }
        },
        "get": {
            "summary": "Obtener todas las reservas",
            "operationId": "obtenerTodasLasReservas",
            "responses": {
                "200": {
                    "description": "Lista de reservas",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Reserva"
                                }
                            }
                        }
                    }
                },
                "500": {
                    "description": "Error en el servidor"
                }
            }
        }
    },
    "/reservas/{id}": {
        "get": {
            "summary": "Obtener una reserva específica",
            "operationId": "obtenerReservaPorId",
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "ID de la reserva",
                    "schema": {
                        "type": "string"
                    }
                }
            ],
            "responses": {
                "200": {
                    "description": "Reserva encontrada",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Reserva"
                            }
                        }
                    }
                },
                "404": {
                    "description": "Reserva no encontrada"
                },
                "500": {
                    "description": "Error en el servidor"
                }
            }
        },
        "put": {
            "summary": "Actualizar información de una reserva",
            "operationId": "actualizarReserva",
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "ID de la reserva",
                    "schema": {
                        "type": "string"
                    }
                }
            ],
            "requestBody": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Reserva"
                        }
                    }
                }
            },
            "responses": {
                "200": {
                    "description": "Reserva actualizada",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Reserva"
                            }
                        }
                    }
                },
                "404": {
                    "description": "Reserva no encontrada"
                },
                "400": {
                    "description": "Error en la actualización"
                }
            }
        },
        "delete": {
            "summary": "Eliminar una reserva específica",
            "operationId": "eliminarReserva",
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "ID de la reserva",
                    "schema": {
                        "type": "string"
                    }
                }
            ],
            "responses": {
                "204": {
                    "description": "Reserva eliminada"
                },
                "404": {
                    "description": "Reserva no encontrada"
                },
                "500": {
                    "description": "Error en el servidor"
                }
            }
        }
    },
    "/reservas/search": {
        "get": {
            "summary": "Buscar reservas por criterios",
            "operationId": "buscarReservas",
            "parameters": [
                {
                    "name": "hotel",
                    "in": "query",
                    "required": false,
                    "description": "Nombre del hotel",
                    "schema": {
                        "type": "string"
                    }
                },
                {
                    "name": "fechaEntrada",
                    "in": "query",
                    "required": false,
                    "description": "Fecha de entrada en formato YYYY-MM-DD",
                    "schema": {
                        "type": "string",
                        "format": "date"
                    }
                },
                {
                    "name": "fechaSalida",
                    "in": "query",
                    "required": false,
                    "description": "Fecha de salida en formato YYYY-MM-DD",
                    "schema": {
                        "type": "string",
                        "format": "date"
                    }
                },
                {
                    "name": "tipoHabitacion",
                    "in": "query",
                    "required": false,
                    "description": "Tipo de habitación",
                    "schema": {
                        "type": "string"
                    }
                }
            ],
            "responses": {
                "200": {
                    "description": "Lista de reservas que coinciden con los criterios de búsqueda",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Reserva"
                                }
                            }
                        }
                    }
                },
                "404": {
                    "description": "No se encontraron reservas que coincidan con los criterios"
                },
                "500": {
                    "description": "Error en el servidor"
                }
            }
        }
    }
}